requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
pytest
pytest-mock
moto
boto3
pylint